+++ rasa_core/constants.py
@@ -8,5 +8,6 @@ DEFAULT_SERVER_PORT = 5005
 
 DEFAULT_SERVER_URL = "http://localhost:{}".format(DEFAULT_SERVER_PORT)
 
-
 MINIMUM_COMPATIBLE_VERSION = "0.10.0a3"
+
+DOCS_BASE_URL = "https://rasa.com/docs/core"